requirements.txt
numpy~=1.19.5
opencv-python~=4.2.0.34
tensorflow-gpu==1.14.0
pyserial==3.5
six~=1.15.0
pandas~=1.1.5
joblib~=1.0.1
scikit-learn~=0.24.1
requests~=2.0.0
rectpack~=0.2.1
Kivy~=2.0.0